Working on ubiging addition.

+++ code/ubigint.cpp
@@ -38,13 +38,13 @@ ubigint ubigint::operator+ (const ubigint& that) const {
     int digit2;
     int carry = 0;
     for (auto it = that.ubig_value.rbegin(); it != that.ubig_value.rend(); it++) {
-    	digit2 = 0;
-    	digit = it;
-    	for (auto it2 = ubig_value.rbegin(); it != ubig_value.rend(); it2++) {
-    		digit2 = it2;
-    	}
-    	cout << "digit: " << digit;
-    	cout << "digit2: " << digit2;
+        digit2 = 0;
+        digit = static_cast<int> it;
+        for (auto it2 = ubig_value.rbegin(); it != ubig_value.rend(); it2++) {
+            digit2 = static_cast<int> it2;
+        }
+        cout << "digit: " << digit;
+        cout << "digit2: " << digit2;
     }
     return ubigint (ubig_value[0] + that.ubig_value[0]);
 }
